Drop 'setuptools' dependency in setup.py
I can't find a reason to have this runtime dependency, and it isn't
present in stdeb.cfg anyway.

Also move the 'install_requires' with the other setup data to simplify
the setup.py. This was present to facilitate the conditional addition of
'argparse' for Python <= 2.6, which was dropped in a previous change.