Avoid an assert when running into entries that have no release version.
A relatively new standard operating practice in the distro
files in https://github.com/ros/rosdistro is to leave released
packages in place, but to remove their version numbers for
known-failing versions.  Unfortunately, this was tripping up
the assert in _generate_rosinstall_for_package, since that
assumed that the release repository entry had a version.

Fix it here by causing repositories setup this way to behave
like "unreleased" repositories, which I think makes sense.
That avoids the assert and allows it to succeed.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>